Complete upgrade to rust workspaces.
