+++ rs-parser/parser/src/instruction/mov.rs
@@ -1,8 +1,8 @@
 use nom::IResult;
-use nom::character::complete::space0;
+// use nom::character::complete::space0;
 
-use instruction::Operation;
-use instruction::base::*;
+use crate::instruction::Operation;
+// use crate::instruction::base::*;
 
 fn mov_from_in(input: &[u8]) -> IResult<&[u8], Operation> {
   todo!()
@@ -45,9 +45,9 @@ pub fn mov_operation(input: &[u8]) -> IResult<&[u8], Operation> {
 mod tests {
   use super::*;
 
-	use common::to_input;
-  use common::tests::*;
-  use instruction::ValuePointer;
+	use crate::common::to_input;
+  use crate::common::tests::*;
+  use crate::instruction::ValuePointer;
 
   #[test]
   fn test_parse_mov_in_to_out() {
